Improve the stack trace formatting
